Fix authorization for bitcoin average api
Date.now() previously had not worked because it is more
precise than the api's timestamp. With conversion and
rounding it works correctly.

The digestValue must be encoded in hex value first.

Unauthorized requests to the Bitcoin Average API face a 700
request limit per 24 horus. This limit was hit in testing, and
so correct authorization is required to ensure consistent
access to the api.